feat: add tests for adapter_ha pooled connector and refactor related logic
