+++ nodejs/test/tmq/tmq.test.ts
@@ -1,4 +1,3 @@
-import { ClusterRegistry } from "@src/client/clusterRegistry";
 import { TMQConstants } from "@src/tmq/constant";
 import { WsConsumer } from "@src/tmq/wsTmq";
 import { WSConfig } from "@src/common/config";
@@ -19,10 +18,6 @@ let dropTopic = `DROP TOPIC IF EXISTS ${topics[0]};`;
 
 let dsn = `ws://${testUsername()}:${testPassword()}@localhost:6041`;
 let tmqDsn = "ws://localhost:6041";
-
-function resetClusterRegistrySingleton(): void {
-    (ClusterRegistry as any)._instance = undefined;
-}
 
 beforeAll(async () => {
     let conf: WSConfig = new WSConfig(dsn);
@@ -398,95 +393,6 @@ describe("TDWebSocket.Tmq()", () => {
 
         await consumer.unsubscribe();
         await consumer.close();
-    });
-
-    test("adapter_ha consumes data and reuses pooled connector across seed variants", async () => {
-        const now = Date.now();
-        const dsnA = "ws://localhost:6041?adapter_ha=true";
-        const dsnB = "ws://localhost:6041,127.0.0.1:6041?adapter_ha=true";
-        const buildConfig = (
-            groupId: string,
-            clientId: string,
-            url: string
-        ): Map<string, any> => {
-            const cfg = new Map<string, any>();
-            cfg.set(TMQConstants.GROUP_ID, groupId);
-            cfg.set(TMQConstants.CONNECT_USER, testUsername());
-            cfg.set(TMQConstants.CONNECT_PASS, testPassword());
-            cfg.set(TMQConstants.AUTO_OFFSET_RESET, "earliest");
-            cfg.set(TMQConstants.CLIENT_ID, clientId);
-            cfg.set(TMQConstants.WS_URL, url);
-            cfg.set(TMQConstants.ENABLE_AUTO_COMMIT, false);
-            cfg.set(TMQConstants.AUTO_COMMIT_INTERVAL_MS, 1000);
-            cfg.set("session.timeout.ms", "10000");
-            cfg.set("max.poll.interval.ms", "30000");
-            cfg.set("msg.with.table.name", "true");
-            return cfg;
-        };
-
-        let consumerA: WsConsumer | null = null;
-        let consumerB: WsConsumer | null = null;
-
-        WebSocketConnectionPool.instance().destroyed();
-        resetClusterRegistrySingleton();
-
-        try {
-            const cfgA = buildConfig(
-                `g_adapter_ha_${now}_a`,
-                `c_adapter_ha_${now}_a`,
-                dsnA
-            );
-            consumerA = await WsConsumer.newConsumer(cfgA);
-            await consumerA.subscribe(topics);
-
-            const connectorA = ((consumerA as any)._wsClient as any)._wsConnector;
-            expect(connectorA).toBeDefined();
-            expect(connectorA.getPoolKey()).toMatch(/^ws:\/\/[0-9a-f-]{36}\/rest\/tmq#auth=/);
-
-            let count = 0;
-            for (let i = 0; i < 20 && count < 10; i++) {
-                const res = await consumerA.poll(500);
-                for (const [, value] of res) {
-                    const data = value.getData();
-                    if (!data || data.length === 0) {
-                        continue;
-                    }
-                    count += data.length;
-                }
-            }
-            expect(count).toBeGreaterThanOrEqual(10);
-
-            await consumerA.unsubscribe();
-            await consumerA.close();
-            consumerA = null;
-
-            const cfgB = buildConfig(
-                `g_adapter_ha_${now}_b`,
-                `c_adapter_ha_${now}_b`,
-                dsnB
-            );
-            consumerB = await WsConsumer.newConsumer(cfgB);
-
-            const connectorB = ((consumerB as any)._wsClient as any)._wsConnector;
-            expect(connectorB).toBe(connectorA);
-
-            await consumerB.subscribe(topics);
-            const res = await consumerB.poll(500);
-            expect(res).toBeTruthy();
-
-            await consumerB.unsubscribe();
-            await consumerB.close();
-            consumerB = null;
-        } finally {
-            if (consumerB) {
-                await consumerB.close();
-            }
-            if (consumerA) {
-                await consumerA.close();
-            }
-            WebSocketConnectionPool.instance().destroyed();
-            resetClusterRegistrySingleton();
-        }
     });
 
     testEnterprise("connect with token", async () => {
